Move healthcheck handler to http package
