+++ todo.go
@@ -7,25 +7,8 @@ import (
 
 	"github.com/go-chi/chi"
 	"github.com/go-chi/chi/middleware"
+	todoHttp "github.com/overkilling/overkill-todo-monolith-api/http"
 )
-
-type serviceUpCheck func() bool
-
-func healthcheckHandler(dbServiceAlive serviceUpCheck) http.HandlerFunc {
-	type HealthcheckResponse struct {
-		Status string `json:"status"`
-	}
-	return func(w http.ResponseWriter, r *http.Request) {
-		status := "fail"
-		if dbServiceAlive() {
-			status = "ok"
-		}
-		response := HealthcheckResponse{status}
-		responseBytes, _ := json.Marshal(response)
-		w.Header().Set("Content-Type", "application/json")
-		w.Write(responseBytes)
-	}
-}
 
 func getTodosHandler() http.HandlerFunc {
 	type Todo struct {
@@ -48,7 +31,7 @@ func Router(db *sql.DB) http.Handler {
 	r.Use(middleware.Recoverer)
 	r.Use(middleware.Heartbeat("/ping"))
 
-	r.Get("/health", healthcheckHandler(func() bool { return db.Ping() == nil }))
+	r.Get("/health", todoHttp.NewHealthcheckHandler(func() bool { return db.Ping() == nil }))
 	r.Get("/todos", getTodosHandler())
 
 	return r
